+++ Makefile
@@ -6,7 +6,7 @@
 #    By: nachab <nachab@student.1337.ma>            +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/11/25 17:00:59 by nachab            #+#    #+#              #
-#    Updated: 2023/11/28 15:59:54 by nachab           ###   ########.fr        #
+#    Updated: 2023/11/29 12:35:28 by nachab           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -20,7 +20,7 @@ FRAMEWORKS = /Users/nachab/Desktop/CUB3D_42/MLX/libmlx42.a -Iinclude -lglfw -L $
 
 FLAGS = -Wall -Wextra -Werror #-g -fsanitize=address
 
-PARSE_SRC =  parsing/free_ressources.c parsing/rgb_conversion.c parsing/get_file_data.c parsing/init_data.c parsing/store_data.c parsing/check_map_utils.c parsing/check_map_utils2.c parsing/file_parser.c parsing/check_player.c parsing/store_data_utils.c
+PARSE_SRC =  parsing/texel.c parsing/free_ressources.c parsing/rgb_conversion.c parsing/get_file_data.c parsing/init_data.c parsing/store_data.c parsing/check_map_utils.c parsing/check_map_utils2.c parsing/file_parser.c parsing/check_player.c parsing/store_data_utils.c
 
 GNL_SRC = get_next_line/get_next_line.c get_next_line/get_next_line_utils.c
